Refactor ContentView into HomeView and centralize navigation
