+++ pokedexSwiftUI/pokedexSwiftUIApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct pokedexSwiftUIApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            AppNavigationView()
         }
     }
 }
